requirements.txt
Django==4.2.7
# mysqlclient==2.2.0  # 在 Windows 上可能需要编译，使用 pymysql 作为替代
pymysql==1.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.0.0  # 圖片處理庫，用於 ImageField（Windows 用戶建議使用預編譯版本）
